+++ src/mec.constraint.js
@@ -519,18 +519,18 @@ mec.constraint = {
                             .drw({d:mec.constraint.arrow[type],lsh:true})
                           .end();
 
-            if (mec.showConstraintLabels) {
+            if (this.model.labels.constraints) {
                 let idstr = id || '?', cw = Math.cos(w), sw = Math.sin(w),
                       xid = p1.x + 20*cw - 10*sw, 
                       yid = p1.y + 20*sw + 10*cw;
-                if (this.ori.type === 'ref') {
-                    idstr += '('+ this.ori.ref.id+')';
+                if (this.ori.type === 'ref' || this.len.type === 'ref') {
+                    idstr += '('+ (this.ori.type === 'ref' ? this.ori.ref.id : this.len.ref.id) +')';
                     xid -= 3*sw;
                     yid += 3*cw;
-                }  
+                }
                 g.txt({str:idstr,x:xid,y:yid,thal:'center',tval:'middle', ls:mec.txtColor})
             }
-            
+
             return g;
         }
     },