+++ src/mec.load.js
@@ -124,7 +124,7 @@ mec.load.force = {
                             lc:'round',sh:()=>this.sh,fs:'@ls'})
                       .drw({d:mec.load.force.arrow,lsh:true})
                       .end();
-        if (mec.showLoadLabels)
+        if (this.model.labels.loads)
             g.txt({str:this.id||'?',x:xid,y:yid,thal:'center',tval:'middle',ls:mec.txtColor});
         return g;
     },