+++ changes.md
@@ -203,7 +203,10 @@ depending on its `shape.type` shapes become one of these objects:
 ```
 
 # Changes 11/08/18
-
+* add `stroke` property to `mec.trace` elements.
+* changed source to current `mec.labels` convention.
+* introduce `mec.labels` structure with previous defaults for setting per JSON model.
+* depricate `mec.showNodeLabels`, `mec.showConstraintLabels` , `mec.showLoadLabels`.
 * decouple element prototype extension and initialization.
 * static `mec.aly.*` for processing `view.info` added.
 * `view.info` added.